+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -6,10 +6,11 @@ CHANGELOG
 
  * added `StreamWrapper`
  * added `HttplugClient`
+ * added `max_duration` option
  * added support for NTLM authentication
  * added `$response->toStream()` to cast responses to regular PHP streams
  * made `Psr18Client` implement relevant PSR-17 factories and have streaming responses
- * added `max_duration` option
+ * added `TraceableHttpClient`, `HttpClientDataCollector` and `HttpClientPass` to integrate with the web profiler
 
 4.3.0
 -----